+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+#### Features
+
+- Support for Distributed Tracing ([957](https://github.com/getsentry/sentry-elixir/pull/957))
+
 ## 11.0.4
 
 - Fix safe JSON encoding of improper lists ([#938](https://github.com/getsentry/sentry-elixir/pull/938))